Updated login and chat style to match
