+++ public/js/login.js
@@ -19,9 +19,35 @@ const loginFormHandler = async (event) => {
   }
 };
 
+const signupFormHandler = async (event) => {
+  event.preventDefault();
+
+  const username = document.querySelector('#username-signup').value.trim();
+  const email = document.querySelector('#email-signup').value.trim();
+  const password = document.querySelector('#password-signup').value.trim();
+
+  if (username && email && password) {
+    const response = await fetch('/api/users', {
+      method: 'POST',
+      body: JSON.stringify({ username, email, password }),
+      headers: { 'Content-Type': 'application/json' },
+    });
+
+    if (response.ok) {
+      document.location.replace('/');
+    } else {
+      alert('Failed to sign up.');
+    }
+  }
+};
+
+
+
 document
   .querySelector('.login-form')
   .addEventListener('submit', loginFormHandler);
 
-
+document
+  .querySelector('.signup-form')
+  .addEventListener('submit', signupFormHandler);
