+++ views/chat.handlebars
@@ -1,7 +1,9 @@
 <body>
   <div class="messageBox">
-    <ul id="messages"></ul>
-    Testing
+    <div class="innerBox">
+      <ul id="messages"></ul>
+      username: Testing
+    </div>
   </div>
 
   <div class="input-group mb-3 messageInput">
