+++ views/layouts/main.handlebars
@@ -6,7 +6,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
   <link href='https://fonts.googleapis.com/css?family=Baloo+2:400,800&display=swap' rel='stylesheet'/>
-  <link rel="stylesheet" href="/css/jass.css">
   <link rel="stylesheet" href="/css/style.css">
   <title>Cumulus Chat</title>
 </head>
@@ -25,7 +24,7 @@
         <a href="/chat">chat</a> |
         <button class="no-button" id="logout" href='/login'>logout</button>
         {{else}}
-        {{{startcloud}}}
+        
         {{/if}}
       </nav>
     </header>
@@ -42,10 +41,7 @@
 
   <!-- Render script for logged in users only -->
   {{#if logged_in}}
-  {{!-- <script src="../../public/js/logout.js"></script> --}}
-  {{else}} 
   <script src="/public/js/logout.js"></script>
-  <script src='/public/js/signup.js'></script>
   {{/if}}
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
