+++ views/login.handlebars
@@ -29,8 +29,9 @@
                 />
               </div>
               <h6><a href='#!'>Forgot password?</a></h6>
+              <br>
 
-              <button type='submit'>Login</button>
+              <button class="flowButton" type='submit'>Login</button>
             </div>
             <div>
               <h6>Don't have an account?</h6>
@@ -62,7 +63,7 @@
                     />
                   </div>
 
-                  <button type='submit'>Sign Up</button>
+                  <button class="flowButton" type='submit'>Sign Up</button>
                 </div>
             </div>
           </div>
@@ -70,5 +71,5 @@
       </div>
     </div>
   </div>
-<script src="./public/js/login.js"></script>
 </section>
+<script src="./js/login.js"></script>
