+++ views/startcloud.handlebars
@@ -1,5 +1,3 @@
-<div class='container py-5 h-100'>
-    <div class='row d-flex justify-content-center align-items-center h-100'>
-        <a href='/login' class='text-black-50 fw-bold center'><button type='submit'>START</button></a>
-    </div>
+<div class='centerDiv'>
+        <a href='/login' class='center'><button type='submit'>START</button></a>
 </div>
